section 1.4: HomeViewModel & HomeView Refactor
