+++ FitnessApp/Home/Models/Workout.swift
@@ -1,0 +1,17 @@
+//
+//  Workout.swift
+//  FitnessApp
+//
+//  Created by Jason Dubon on 7/31/23.
+//
+
+import SwiftUI
+struct Workout {
+    let id: Int
+    let title: String
+    let image: String
+    let tintColor: Color
+    let duration: String
+    let date: String
+    let calories: String
+}
